+++ examples/hello_asyncio.py
@@ -38,6 +38,7 @@ import stat
 import logging
 import errno
 import pyfuse3
+import pyfuse3_asyncio
 
 try:
     import faulthandler
@@ -47,6 +48,7 @@ else:
     faulthandler.enable()
 
 log = logging.getLogger(__name__)
+pyfuse3_asyncio.enable()
 
 class TestFs(pyfuse3.Operations):
     def __init__(self):
@@ -146,7 +148,7 @@ def main():
     pyfuse3.init(testfs, options.mountpoint, fuse_options)
     loop = asyncio.get_event_loop()
     try:
-        loop.run_until_complete(pyfuse3.main(aio='asyncio'))
+        loop.run_until_complete(pyfuse3.main())
     except:
         pyfuse3.close(unmount=False)
         raise